separates utility function in csv_importer
